"Stacked" events now work, such as alarm and keyboard. Default step and draw code persists. Local lists are loaded from various YAML files under ENIGMAsystem/Extensions. The Game.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@311 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/GSsprite.h
@@ -35,7 +35,7 @@ int draw_sprite_stretched(int spr,int subimg,double x,double y,double w,double h
 //enigma::4args int draw_sprite_tiled(ARG sprite,ARG2 subimg,ARG3 x,ARG4 y)
 int draw_sprite_part(int spr,int subimg,double left,double top,double width,double height,double x,double y);
 int draw_sprite_part_offset(int spr,int subimg,double left,double top,double width,double height,double x,double y);
-int draw_sprite_ext(int spr,int subimg,double x,double y,double xscale,double yscale,double rot,double color,double alpha);
+int draw_sprite_ext(int spr,int subimg,double x,double y,double xscale,double yscale,double rot,int color,double alpha);
 //draw_sprite_stretched_ext(sprite,subimg,x,y,w,h,color,alpha)
 //draw_sprite_tiled_ext(sprite,subimg,x,y,xscale,yscale,color,alpha)
 int draw_sprite_part_ext(int spr,int subimg,double left,double top,double width,double height,double x,double y,double xscale,double yscale,int color,double alpha);
